+++ src/GoingMy.UserService/src/GoingMy.User.Infrastructure/Migrations/20260429093808_AddUserBlocks.cs
@@ -1,0 +1,45 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace GoingMy.User.Infrastructure.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddUserBlocks : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.CreateTable(
+                name: "UserBlocks",
+                columns: table => new
+                {
+                    BlockerId = table.Column<Guid>(type: "uuid", nullable: false),
+                    BlockeeId = table.Column<Guid>(type: "uuid", nullable: false),
+                    CreatedAt = table.Column<DateTime>(type: "timestamp with time zone", nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_UserBlocks", x => new { x.BlockerId, x.BlockeeId });
+                });
+
+            migrationBuilder.CreateIndex(
+                name: "IX_UserBlocks_BlockeeId",
+                table: "UserBlocks",
+                column: "BlockeeId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_UserBlocks_BlockerId",
+                table: "UserBlocks",
+                column: "BlockerId");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropTable(
+                name: "UserBlocks");
+        }
+    }
+}